+++ src/main/java/dev/arkav/openoryx/net/listeners/Invokeable.java
@@ -1,0 +1,14 @@
+package dev.arkav.openoryx.net.listeners;
+
+import dev.arkav.openoryx.net.data.Packet;
+import dev.arkav.openoryx.net.packets.PacketType;
+
+import java.io.IOException;
+
+public interface Invokeable {
+    void invoke(Packet p) throws IOException;
+    void invoke(ListenerType l) throws Exception;
+    default boolean has(PacketType type) {
+        return true;
+    }
+}