+++ changelog.md
@@ -4,3 +4,6 @@
  - Deprecate mapping objects/packets from resources.
  - Update URLs in update.bat
  - Add latest client token.
+##1.2.0
+ - PacketIO and ProxyIO now call to a Invokable interface
+ - Refactored endpoint building